requirements.txt
# Web Framework
fastapi==0.115.12
uvicorn[standard]==0.34.2

# Data Models & Settings
pydantic==2.10.3
pydantic-settings==2.9.1

# Environment Management
python-dotenv==1.0.1

# LLM & AI Services
openai==1.60.2
google-genai==1.32.0

# Vector Database
qdrant-client==1.14.2

# Document Database  
pymongo==4.8.0

# Observability & Monitoring
langfuse==2.60.5

# Retry & Error Handling
tenacity==8.5.0

# HTTP Client
httpx==0.28.1

# PDF Processing
PyMuPDF==1.26.1
Pillow>=10.0.0

# Additional utilities
python-multipart==0.0.20
pyyaml==6.0.2
aiofiles==24.1.0

# Testing
pytest==8.4.1
pytest-asyncio==0.25.0